Consolidate rectangle tests.
